+++ src/Sparrow/Sparrow/Resources/Dftb/ParameterSets.cpp
@@ -7,9 +7,11 @@
 #include "Sparrow/Resources/Dftb/ParameterSets.h"
 #include "Sparrow/Resources/Dftb/3ob-2-1/Parameters.h"
 #include "Sparrow/Resources/Dftb/3ob-3-1/Parameters.h"
+#include "Sparrow/Resources/Dftb/borg-0-1/Parameters.h"
 #include "Sparrow/Resources/Dftb/hyb-0-2/Parameters.h"
 #include "Sparrow/Resources/Dftb/mio-0-1/Parameters.h"
 #include "Sparrow/Resources/Dftb/mio-1-1/Parameters.h"
+#include "Sparrow/Resources/Dftb/trans3d-0-1/Parameters.h"
 #include "Sparrow/Resources/Dftb/znorg-0-1/Parameters.h"
 
 namespace Scine {
@@ -39,6 +41,14 @@ boost::optional<ParameterSet> embeddedParameters(const std::string& name, const 
 
   if (name == "hyb-0-2") {
     return params_mio_0_1(elements).patch(params_hyb_0_2(elements));
+  }
+
+  if (name == "trans3d-0-1") {
+    return params_mio_1_1(elements).patch(params_trans3d_0_1(elements));
+  }
+
+  if (name == "borg-0-1") {
+    return params_mio_1_1(elements).patch(params_borg_0_1(elements));
   }
 
   return boost::none;